backend/requirements.txt
PyJWT
opencv-python
pillow
numpy
fastapi[standard]
passlib
python-multipart
pydantic
google-api-python-client
google-generativeai
pandas
matplotlib
pillow-avif-plugin
paddlepaddle
setuptools
paddleocr>=2.0.1
python-dotenv
